feat: decoupled oid4vc service from credo service
