+++ src/app.module.ts
@@ -7,6 +7,9 @@ import { QrcodeModule } from "./qrcode/qrcode.module";
 import { MessageModule } from "./message/message.module";
 import { LedgerModule } from "./ledger/ledger.module";
 import { ConnectionsModule } from "./connections/connections.module";
+import { Oid4vcController } from './oid4vc/oid4vc.controller';
+import { Oid4vcModule } from './oid4vc/oid4vc.module';
+import { Oid4vcService } from "./oid4vc/oid4vc.service";
 
 @Module({
   imports: [
@@ -16,8 +19,9 @@ import { ConnectionsModule } from "./connections/connections.module";
     MessageModule,
     LedgerModule,
     ConnectionsModule,
+    Oid4vcModule,
   ],
-  controllers: [AppController],
-  providers: [AppService],
+  controllers: [AppController, Oid4vcController],
+  providers: [AppService, Oid4vcService],
 })
 export class AppModule {}
